requirements.txt
Flask
h5py
gunicorn
numpy
scipy
Keras
Pillow
opencv-python
Flask-WTF
efficientnet